add check access to btn and manage tools

+++ src/component/book/BookManage/BookManage.tsx
@@ -23,6 +23,7 @@ import 'moment/locale/ar';
 import moment from 'moment';
 import moment_jalaali from 'moment-jalaali';
 import { FixNumber } from "../../form/fix-number/FixNumber";
+import { AccessService } from "../../../service/service.access"
 
 /// define props & state ///////
 export interface IProps {
@@ -165,7 +166,7 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
             let minute;
             let second;
             if (row.duration) {
-              if(row.duration === 'NaN' ){
+              if (row.duration === 'NaN') {
                 return ''
               }
               let totalTime = Number(row.duration);
@@ -213,23 +214,29 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
           }
         },
       ],
-      actions: [
+
+      actions: this.checkAllAccess() ? [
         {
+          access : (row : any) => {return this.checkDeleteToolAccess()},
           text: <i title={Localization.remove} className="fa fa-trash text-danger"></i>,
           ac_func: (row: any) => { this.onShowRemoveModal(row) },
           name: Localization.remove
         },
         {
+          access : (row : any) => {return this.checkUpdateToolAccess()},
           text: <i title={Localization.update} className="fa fa-pencil-square-o text-primary"></i>,
           ac_func: (row: any) => { this.updateRow(row) },
           name: Localization.update
         },
         {
+          access : (row : any) => {return this.checkPriceAddToolAccess()},
           text: <i title={Localization.Pricing} className="fa fa-money text-success"></i>,
           ac_func: (row: any) => { this.onShowPriceModal(row) },
           name: Localization.Pricing
         },
       ]
+      :
+      undefined
     },
     BookError: undefined,
     pager_offset: 0,
@@ -263,6 +270,36 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
   private _bookService = new BookService();
   private _priceService = new PriceService();
 
+  checkAllAccess():boolean{
+    if(AccessService.checkOneOFAllAccess(['BOOK_DELETE_PREMIUM','BOOK_EDIT_PREMIUM','PRICE_ADD_PREMIUM'])
+    || AccessService.checkOneOFAllAccess(['BOOK_DELETE_PRESS','BOOK_EDIT_PRESS','PRICE_ADD_PRESS'])
+    ){
+      return true;
+    }
+    return false;
+  }
+
+  checkDeleteToolAccess():boolean{
+    if(AccessService.checkAccess('BOOK_DELETE_PREMIUM') || AccessService.checkAccess('BOOK_DELETE_PRESS')){
+      return true;
+    }
+    return false
+  }
+
+  checkUpdateToolAccess():boolean{
+    if(AccessService.checkAccess('BOOK_EDIT_PREMIUM') || AccessService.checkAccess('BOOK_EDIT_PRESS')){
+      return true;
+    }
+    return false
+  }
+
+  checkPriceAddToolAccess():boolean{
+    if(AccessService.checkAccess('PRICE_ADD_PREMIUM') || AccessService.checkAccess('PRICE_ADD_PRESS')){
+      return true;
+    }
+    return false
+  }
+
   // constructor(props: IProps) {
   //   super(props);
   //   // this._bookService.setToken(this.props.token)
@@ -279,6 +316,9 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
   }
 
   updateRow(book_id: any) {
+    if(this.checkUpdateToolAccess() === false){
+      return;
+    }
     this.props.history.push(`/book/${book_id.id}/edit`);
   }
 
@@ -305,6 +345,9 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
   // delete modal function define
 
   onShowRemoveModal(book: IBook) {
+    if(this.checkDeleteToolAccess() === false ){
+      return;
+    }
     this.selectedBook = book;
     this.setState({ ...this.state, removeModalShow: true });
   }
@@ -315,6 +358,9 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
   }
 
   async onRemoveBook(book_id: string) {
+    if(this.checkDeleteToolAccess() === false ){
+      return;
+    };
     this.setState({ ...this.state, setRemoveLoader: true });
     let res = await this._bookService.remove(book_id).catch(error => {
       this.handleError({ error: error.response });
@@ -358,6 +404,9 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
   }
 
   onShowPriceModal(book: IBook) {
+    if(this.checkPriceAddToolAccess() === false){
+      return;
+    }
     this.selectedBook = book;
     this.setState({
       ...this.state, priceModalShow: true, price: {
@@ -371,6 +420,9 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
     this.setState({ ...this.state, priceModalShow: false });
   }
   async onPriceBook(book_id: string) {
+    if(this.checkPriceAddToolAccess() === false){
+      return;
+    };
     if (!this.state.price.isValid) return;
     this.setState({ ...this.state, setPriceLoader: true });
     let res = await this._priceService.price(book_id, this.state.price.value!).catch(error => {
@@ -575,6 +627,9 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
   ///// navigation function //////
 
   gotoBookCreate() {
+    if(AccessService.checkAccess('BOOK_ADD_PREMIUM')===false && AccessService.checkAccess('BOOK_ADD_PRESS')===false){
+      return;
+    };
     this.props.history.push('/book/create'); // /admin
   }
 
@@ -708,9 +763,12 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
     return (
       <>
         <div className="content">
-          <div className="row">
-            <div className="col-12">
-              <h2 className="text-bold text-dark pl-3">{Localization.book}</h2>
+          {
+            AccessService.checkAccess('BOOK_ADD_PREMIUM') || AccessService.checkAccess('BOOK_ADD_PRESS')
+              ?
+              <div className="row">
+                <div className="col-12">
+                  <h2 className="text-bold text-dark pl-3">{Localization.book}</h2>
                   <BtnLoader
                     loading={false}
                     disabled={false}
@@ -719,8 +777,11 @@ class BookManageComponent extends BaseComponent<IProps, IState>{
                   >
                     {Localization.new}
                   </BtnLoader>
-            </div>
-          </div>
+                </div>
+              </div>
+              :
+              undefined
+          }
           <div className="row">
             <div className="col-12">
               <div className="template-box mb-4">
